Move things acording to the new dir structure
