+++ test/CMakeLists.txt
@@ -29,6 +29,7 @@ set(tweedledum_tests_files
     ${CMAKE_CURRENT_SOURCE_DIR}/Synthesis/pkrm_synth.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Synthesis/pprm_synth.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Synthesis/spectrum_synth.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/Synthesis/steiner_gauss_synth.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Synthesis/transform_synth.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Synthesis/xag_synth.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Utils/BMatrix.cpp
